Real-time progress in Task Center with block-character bars
Task Center and bottom TaskBar now poll the shared TaskQueue at 10 Hz on
the main event loop instead of subscribing to queue callbacks. The old
subscriber chain dropped >80% of progress updates on bursty tasks
because call_from_thread swallowed exceptions when the target screen
was unmounted. Polling is driven by Textual's own set_interval timer so
every update lands on the render thread regardless of worker cadence.

Each task row renders a Rich Text block-character bar
(████████░░░░░░░░░░░░  45.0%) via a new progress_cell module that
supports determinate percentages and an indeterminate pulse window.
Rows reconcile in place via DataTable.update_cell keyed on task_id
instead of clear+rebuild, so there is no flicker when a burst of
updates arrives.

Setup wizard routes each model download through TaskBarController so
the shared queue reflects what the wizard shows locally. The wizard's
own per-row ProgressBar UI is kept for the install flow; wiring both
was kept deliberately so the install experience is unchanged.

Other fixes bundled here:

- Setup wizard focuses the chat-model grid on mount so arrow keys and
  Enter work without a mouse.
- GridSelect Tab/Shift-Tab always escape the grid so Tab on a card does
  not move the cursor destructively within the grid; within-grid
  navigation now only uses arrow keys.
- DownloadProgress.percent is a float so multi-GB downloads render
  smooth fractional movement instead of integer jumps.
- huggingface_hub DOWNLOAD_CHUNK_SIZE is shrunk from 10 MB to 200 KB
  at import time; the default produced ~7 s gaps between callbacks on
  typical home-internet speeds.
- HF_HUB_DISABLE_XET=1 so the xet transfer layer does not short-circuit
  the tqdm callback cadence (coarse 3-4-jump progress).

+++ src/lilbee/cli/tui/task_queue.py
@@ -44,7 +44,7 @@ class Task:
     task_type: str
     fn: Callable[[], None]
     status: TaskStatus = TaskStatus.QUEUED
-    progress: int = 0
+    progress: float = 0.0
     detail: str = ""
     indeterminate: bool = False
 
@@ -152,7 +152,7 @@ class TaskQueue:
     def update_task(
         self,
         task_id: str,
-        progress: int,
+        progress: float,
         detail: str = "",
         *,
         indeterminate: bool | None = None,
